scheduled api calls

+++ server.rb
@@ -7,15 +7,14 @@ require 'sinatra/reloader'
 require 'pry'
 require 'httparty'
 # require_relative './config/environments'
-require_relative './lib/connection-yoshie'
+# require_relative './lib/connection-yoshie'
 # require_relative './lib/connection-tim'
-# require_relative './lib/connection-eric'
+require_relative './lib/connection-eric'
 require_relative './lib/methods'
 
 after do
   ActiveRecord::Base.connection.close
 end
-
 
 before do
   content_type :json
@@ -38,6 +37,8 @@ def parseNYTimes(events, cat)
     title = checkEndpoint(event['event_name'])
     description = checkEndpoint(event['web_description'])    
     date = checkEndpoint(event['date_time_description'])
+    venue = checkEndpoint(event['venue_name'])
+    venue_website = checkEndpoint(event['venue_website'])
     street = checkEndpoint(event['street_address'])
     city = checkEndpoint(event['city'])
     state = checkEndpoint(event['state'])
@@ -52,6 +53,8 @@ def parseNYTimes(events, cat)
     title: title,
     description: description,  
     date: date,
+    venue: venue,
+    venue_website: venue_website,
     address: street + "\n" + city + "\n" + state + "\n" + postal_code,
     latitude: latitude,
     longitude: longitude,
@@ -63,16 +66,19 @@ end
 
 def newEvents()
   old_events = Event.all()
-  old_events.delete_all()
+  
+  if Time.now.to_s.split(' ')[0].split('-')[2] > old_events.last.created_at.to_s.split(' ')[0].split('-')[2]
+    old_events.delete_all()
 
-  pop = HTTParty.get('http://api.nytimes.com/svc/events/v2/listings.json?filters=category:Pop&date_range:2014-10-10&api-key=bd9c3678d4d278b91d84b1082d19d548:15:65256769')
-  parseNYTimes(pop, 'music')
+    pop = HTTParty.get('http://api.nytimes.com/svc/events/v2/listings.json?filters=category:Pop&date_range:2014-10-10&api-key=bd9c3678d4d278b91d84b1082d19d548:15:65256769')
+    parseNYTimes(pop, 'music')
 
-  art = HTTParty.get('http://api.nytimes.com/svc/events/v2/listings.json?filters=category:Art&date_range:2014-10-10&api-key=bd9c3678d4d278b91d84b1082d19d548:15:65256769')
-  parseNYTimes(art, 'art')
+    art = HTTParty.get('http://api.nytimes.com/svc/events/v2/listings.json?filters=category:Art&date_range:2014-10-10&api-key=bd9c3678d4d278b91d84b1082d19d548:15:65256769')
+    parseNYTimes(art, 'art')
 
-  theater = HTTParty.get('http://api.nytimes.com/svc/events/v2/listings.json?filters=category:Theater&date_range:2014-10-10&api-key=bd9c3678d4d278b91d84b1082d19d548:15:65256769')
-  parseNYTimes(theater, 'theater')
+    theater = HTTParty.get('http://api.nytimes.com/svc/events/v2/listings.json?filters=category:Theater&date_range:2014-10-10&api-key=bd9c3678d4d278b91d84b1082d19d548:15:65256769')
+    parseNYTimes(theater, 'theater')
+  end
 end
 
 newEvents()
